+++ overpassforge/base.py
@@ -3,7 +3,6 @@ from typing import Iterable, TYPE_CHECKING
 from .filters import (
     Filter,
     BoundingBox,
-    Ids,
     Regex,
     Key,
     Intersection,
@@ -72,14 +71,6 @@ class Statement:
     def __hash__(self) -> int:
         return id(self)
     
-    def __sub__(self, other: Statement) -> Difference:
-        from .statements import Difference
-        return Difference(self, other)
-    
-    def __add__(self, other: Statement) -> Union:
-        from .statements import Union
-        return Union(self, other)
-    
     def out(self, *options: str | tuple[float, float, float, float]) -> Statement:
         """
         Indicate that the result of this statement must be outputed.
@@ -139,59 +130,31 @@ class Statement:
         return f"<{self.__class__.__name__} \'{info}\'>"
 
 
-class QueryStatement(Statement):
-    """Represents a query statement, e.g. node, rel, way... Query statements always
-    return a set that can be used as input to other statements/filters.
-    """
+class Set(Statement):
+    """Represents a set, i.e. a statement that always returns a set of elements."""
 
-    _type_specifier: str = "<Unspecified>"
-    
-    def __init__(self,
-        ids: Iterable[int] | int | None = None,
-        label: str | None = None, *,
-        bounding_box: tuple[float, float, float, float] | None = None,
-        input_set: Statement | None = None,
-        within: Areas | None = None,
-        around: tuple[Statement, float] | Around | None = None,
-        filters: Iterable[Filter] = [],
-        **tags: str
-    ) -> None:
-        
+    def __init__(self, filters: Iterable[Filter] = [], label: str | None = None) -> None:
         super().__init__(label)
 
         self.filters = list(filters)
-
-        if isinstance(input_set, Statement):
-            self.filters.append(Intersection(input_set))
-
-        if isinstance(ids, int):
-            self.filters.append(Ids(ids))
-        elif ids is not None:
-            self.filters.append(Ids(*ids))
-
-        if bounding_box is not None:
-            self.filters.append(BoundingBox(*bounding_box))
-        
-        if within is not None:
-            self.filters.append(Area(within))
-        
-        if isinstance(around, Around):
-            self.filters.append(around)
-        elif around is not None:
-            self.filters.append(Around(around[1], around[0]))
-        
-        for key, value in tags.items():
-            self.filters.append(Key(key) == value)
     
-    def filter(self, *filters: Filter) -> QueryStatement:
-        """Adds filters to the statement/set.
+    def __sub__(self, other: Set) -> Difference:
+        from .statements import Difference
+        return Difference(self, other)
+    
+    def __add__(self, other: Set) -> Union:
+        from .statements import Union
+        return Union(self, other)
+    
+    def filter(self, *filters: Filter) -> Set:
+        """Adds filters to the set.
 
         Args:
             filters: A list of filters.
         """
         return self.__class__(filters=[Intersection(self), *filters])
     
-    def where(self, *keys: str | Regex, **tags: str | Regex) -> QueryStatement:
+    def where(self, *keys: str | Regex, **tags: str | Regex) -> Set:
         """Adds filters "key", ~"key", "key"="value", or "key"~"value".
 
         Args:
@@ -199,14 +162,14 @@ class QueryStatement(Statement):
             tags: List of key="value" tags the elements must have.
         """
 
-        filters: list[Filter] = [Intersection(self)]
+        filters: list[Filter] = []
         for key in keys:
             filters.append(Key(key))
         for key, value in tags.items():
             filters.append(Key(key) == value)
-        return self.__class__(filters=filters)
+        return self.filter(*filters)
     
-    def within(self, area: tuple[float,float,float,float] | BoundingBox | Polygon | Area | Areas) -> QueryStatement:
+    def within(self, area: tuple[float,float,float,float] | BoundingBox | Polygon | Area | Areas) -> Set:
         """Filters the elements that are in the specified area.
 
         Args:
@@ -214,60 +177,60 @@ class QueryStatement(Statement):
                 area filter or set of areas.
         """
         if isinstance(area, Filter):
-            return self.__class__(filters=[Intersection(self), area])
+            return self.filter(area)
         elif isinstance(area, tuple):
-            return self.__class__(filters=[Intersection(self), BoundingBox(*area)])
+            return self.filter(BoundingBox(*area))
         else:
-            return self.__class__(filters=[Intersection(self), Area(area)])
+            return self.filter(Area(area))
     
-    def intersection(self, *others: Statement) -> QueryStatement:
+    def intersection(self, *others: Statement) -> Set:
         """Returns the statement computing the intersection of
         this statement with the others.
         
         Args:
             others: Other statements to intersect with
         """
-        return self.__class__(filters=[Intersection(self, *others)])
+        return self.filter(Intersection(*others))
     
-    def changed_since(self, date: datetime) -> QueryStatement:
+    def changed_since(self, date: datetime) -> Set:
         """Filters the elements that were changed since the specified datetime.
 
         Args:
             date: The lower bound of change dates to consider
         """
-        return self.__class__(filters=[Intersection(self), Newer(date)])
+        return self.filter(Newer(date))
     
-    def changed_between(self, lower: datetime, upper: datetime) -> QueryStatement:
+    def changed_between(self, lower: datetime, upper: datetime) -> Set:
         """Filters the elements that were changed between the two specified dates.
 
         Args:
             lower (datetime): Lower bound datetime.
             upper (datetime): Upper bound datetime.
         """
-        return self.__class__(filters=[Intersection(self), Changed(lower, upper)])
+        return self.filter(Changed(lower, upper))
     
-    def last_changed_by(self, *users: str | int) -> QueryStatement:
+    def last_changed_by(self, *users: str | int) -> Set:
         """Filters the elements that last changed by any of the given users.
 
         Args:
             users: The list of user names or user ids.
         """
-        return self.__class__(filters=[Intersection(self), User(*users)])
+        return self.filter(User(*users))
     
-    def outlines_of(self, area: 'Areas') -> QueryStatement:
+    def outlines_of(self, area: 'Areas') -> Set:
         """Filters the elements that are part of the outline of the given area.
 
         Args:
             area: The area to consider to outline of.
         """
-        return self.__class__(filters=[Intersection(self), Pivot(area)])
+        return self.filter(Pivot(area))
     
     def around(self,
         radius: float,
         other: Statement | None = None,
         lats: Iterable[float] | None = None, 
         lons: Iterable[float] | None = None
-    ):
+    ) -> Set:
         """Filters elements that are within a given radius of the elements of another set
         or a list of given coordinates (cannot specify both).
 
@@ -277,24 +240,4 @@ class QueryStatement(Statement):
             lats: List of latitude of points.
             lons: List of longitudes of points.
         """
-        around = Around(radius, other, lats, lons)
-        return self.__class__(filters=[Intersection(self), around])
-
-    @property
-    def _dependencies(self) -> list[Statement]:
-        deps: list[Statement] = []
-        for filt in self.filters:
-            deps.extend(filt._dependencies)
-        return deps
-    
-    def _compile_core(self, vars: _VariableManager, out_var: str | None = None) -> str:
-        comp_filter = lambda f: f._compile(vars)
-        res = self._type_specifier + "".join(map(comp_filter, self.filters))
-        if out_var is not None:
-            return res + f"->.{out_var};"
-        return res + ";"
-
-
-class BlockStatement(Statement):
-    """Represents a block statement, i.e. a statement which acts on multiple statements."""
-    pass
+        return self.filter(Around(radius, other, lats, lons))